Pool Edits
Added Ball Six and Seven
All balls needed to be added, sprites created but not correctly colored yet!
Collitions for new ball objects needed to be added to pockets, as well as collitions with the cue ball and new balls.
Two new pockets need to be added, or make up some game that uses four pockets.